refactor: replace implicit singletons with DI and decompose monolithic modules
Replace module-level singleton access (get_database(), get_notifier(),
etc.) with constructor-based dependency injection, and break down
monolithic `core.py` and `api.py` into focused, cohesive packages:

- Extract `core.py` into `core/` package: processor, searcher, models,
  registry, injector, and utils modules
- Extract `api.py` into `trackers/` package: api_common, gazelle_games,
  gazelle_html, gazelle_json, and registry modules
- Extract client instantiation from `client_instance.py` into
  `clients/registry.py` with a proper factory pattern
- Introduce dependency injection via constructors for TorrentClient
  subclasses (database, scheduler, notifier)
- Update `cli.py` initialization to use explicit dependency wiring with
  AsyncIOScheduler
- Update `scheduler.py` and `webserver.py` for new module structure
