+++ src/nemorosa/clients/__init__.py
@@ -12,20 +12,28 @@ from .client_common import (
 )
 from .deluge import DelugeClient
 from .qbittorrent import QBittorrentClient
+from .registry import (
+    create_torrent_client,
+    get_torrent_client,
+    init_torrent_client,
+)
 from .rtorrent import RTorrentClient
 from .transmission import TransmissionClient
 
 __all__ = [
     "ClientTorrentFile",
     "ClientTorrentInfo",
+    "DelugeClient",
     "FieldSpec",
     "PostProcessResult",
     "PostProcessStatus",
+    "QBittorrentClient",
+    "RTorrentClient",
     "TorrentClient",
     "TorrentConflictError",
     "TorrentState",
-    "DelugeClient",
-    "QBittorrentClient",
-    "RTorrentClient",
     "TransmissionClient",
+    "create_torrent_client",
+    "get_torrent_client",
+    "init_torrent_client",
 ]
